+++ common-sdk/src/main/kotlin/com/algorand/wallet/swap/data/model/SwapQuoteRequestBody.kt
@@ -29,5 +29,5 @@ internal data class SwapQuoteRequestBody(
     @SerializedName("amount")
     val amount: BigInteger,
     @SerializedName("slippage")
-    val slippage: Float?
+    val slippage: Double?
 )
