+++ common-sdk/src/main/kotlin/com/algorand/wallet/swap/domain/usecase/SwapUseCases.kt
@@ -30,8 +30,8 @@ import com.algorand.wallet.swap.domain.model.SwapSelectedAssetDetail
 import com.algorand.wallet.swap.domain.model.SwapStatusFailureReason
 import com.algorand.wallet.swap.domain.model.TopSwapPairs
 import com.algorand.wallet.transaction.domain.model.TransactionId
-import java.math.BigDecimal
 import kotlinx.coroutines.flow.Flow
+import java.math.BigDecimal
 
 fun interface GetSwapQuotes {
     suspend operator fun invoke(payload: SwapQuotePayload): PeraResult<SwapQuotes>
@@ -102,11 +102,11 @@ fun interface SendSwapTransactions {
 }
 
 fun interface SetSwapSlippageTolerancePercentage {
-    suspend operator fun invoke(percentage: Float?)
+    suspend operator fun invoke(percentage: Double?)
 }
 
 fun interface GetSwapSlippageTolerancePercentage {
-    suspend operator fun invoke(): Float?
+    suspend operator fun invoke(): Double?
 }
 
 internal fun interface GetParsedSwapTransactions {
